fix: fix the useSearchParams() can only be in server side | SCRUM-313
